pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Andrew <cramatsu@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ultralytics = "^8.3.27"
pillow = "^11.0.0"
colorama = "^0.4.6"
onnx = "^1.17.0"
onnxruntime = "^1.20.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
blurryface = "app.main:start"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120